% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{california}
\alias{california}
\title{California proposition 99}
\format{A data frame with 1209 rows and 4 variables:
\describe{
\item{State}{US state name, character string}
\item{Year}{Year, integer}
\item{PacksPerCapita}{per-capita cigarette consumption, numeric}
\item{treated}{the treatmed indicator 0: control, 1: treated, numeric}
}}
\usage{
california()
}
\value{
california_prop99
}
\description{
A dataset containing per-capita cigarette consumption (in packs).
In year 1989 California imposed a Tobacco tax. The column \code{treated} is 1 from then on for California.
}
