% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{ss_vcov}
\alias{ss_vcov}
\title{Calculate Variance-Covariance Matrix for a Fitted Model Object}
\usage{
ss_vcov(estimate, method = "placebo", n_reps)
}
\arguments{
\item{estimate}{A Ssynthdid model}

\item{method}{The CI method. The default is "placebo"}

\item{n_reps}{the number of replicacions c("placebo", "bootstrap")}
}
\description{
Provides variance estimates in the context of staggered adoption based on the following three options.
\itemize{
\item The bootstrap, Algorithm 2 in Arkhangelsky et al.
\item The jackknife, Algorithm 3 in Arkhangelsky et al.
\item Placebo, Algorithm 4 in Arkhangelsky et al.
}
}
