% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthdid.R
\name{ssynth_estimate}
\alias{ssynth_estimate}
\title{Staggered Synthetic Difference-in-Difference Estimation}
\usage{
ssynth_estimate(
  data,
  unit,
  time,
  treated,
  outcome,
  covariates = c(),
  weights_sdid = list(lambda = NULL, omega = NULL),
  update_lambda = is.null(weights_sdid$lambda),
  update_omega = is.null(weights_sdid$omega),
  cov_method = "optimized",
  lambda_intercept = T,
  omega_intercept = T,
  max_iter_pre_sparsify = 100,
  max_iter = 10000,
  sparsify = sparsify_function
)
}
\arguments{
\item{data}{tibble::tibble()}

\item{unit}{character: Column number for variable that identifies each unit}

\item{time}{character: Column number for variable indicating time period.}

\item{treated}{character:  Column number for 0/1 variable. = 1 if unit is
treated in that period}

\item{outcome}{character: Column for outcome variable}

\item{covariates}{vec_character: Columns for covariates}

\item{weights_sdid}{refers original package}

\item{update_lambda}{refers original package}

\item{update_omega}{refers original package}

\item{cov_method}{refers original package}

\item{lambda_intercept}{refers original package}

\item{omega_intercept}{refers original package}

\item{max_iter_pre_sparsify}{refers original package}

\item{max_iter}{refers original package}

\item{sparsify}{refers original package}
}
\value{
Estimate Synthdid for staggered contexts, with a simple covariates
implementation method, returning the information in a list, ATT_staggered,
time-specific information, and others.
}
\description{
Staggered Synthetic Difference-in-Difference Estimation
}
