% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ssynthdid-package.R
\docType{package}
\name{Ssynthdid-package}
\alias{Ssynthdid}
\alias{Ssynthdid-package}
\title{Ssynthdid: Staggered Synthetic Difference-in-Difference Estimation}
\description{
Implementation of Synthetic Difference-in-Differences estimation for staggered treatment adoption.
}
\author{
\strong{Maintainer}: Jhon Flores Rojas \email{fr.jhonk@gmail.com} (\href{https://orcid.org/NONE}{ORCID})

}
\keyword{internal}
