% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ssynthdid_plot}
\alias{ssynthdid_plot}
\title{Plots treated and synthetic control trajectories and overlays a 2x2 diff-in-diff diagram of our estimator.
Requires ggplot2}
\usage{
ssynthdid_plot(estimate)
}
\arguments{
\item{estimate}{A Ssynthdid model}
}
\value{
A list with all the plots for the different treatment periods.
}
\description{
Plots treated and synthetic control trajectories and overlays a 2x2 diff-in-diff diagram of our estimator.
Requires ggplot2
}
