% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ssynthdid_units_plot}
\alias{ssynthdid_units_plot}
\title{Plots unit by unit difference-in-differences. Dot size indicates the weights omega_i
used in the average that yields our treatment effect estimate.}
\usage{
ssynthdid_units_plot(estimate)
}
\arguments{
\item{estimate}{A Ssynthdid model}
}
\value{
A list with all the plots for the different treatment periods.
}
\description{
Plots unit by unit difference-in-differences. Dot size indicates the weights omega_i
used in the average that yields our treatment effect estimate.
}
